Segment_mini_dev/requirement_maruf_env.txt
absl-py==0.9.0
adabelief-pytorch==0.2.0
albumentations==0.4.6
anyio==3.3.4
argon2-cffi==21.1.0
ase==3.19.1
astunparse==1.6.3
attrs==19.3.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.1.5
bokeh==2.3.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.15.0
chainer==7.8.1
chainercv==0.13.1
chardet==3.0.4
colorama==0.4.3
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastprogress==1.0.0
filelock==3.3.1
future==0.18.2
gast==0.3.3
gdown==4.2.0
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googledrivedownloader==0.4
grad-cam==1.3.7
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
immutables==0.16
importlib-metadata==1.7.0
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-git==0.20.0
jupyterlab-server==2.8.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kornia==0.6.3
llvmlite==0.33.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
munch==2.5.0
nbclassic==0.3.4
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.7
networkx==2.4
nodejs==0.1.1
notebook==6.0.3
numba==0.50.1
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
optional-django==0.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
pretrainedmodels==0.7.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDOE==0.3.8
Pygments==2.6.1
pynrrd==0.4.2
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
rdflib==5.0.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.1.0
seaborn==0.11.0
segmentation-models-pytorch==0.1.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
smmap==3.0.4
sniffio==1.2.0
soupsieve==2.2.1
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.4
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.22
torch==1.10.1
torch-cluster==1.5.6
torch-geometric==1.6.0
torch-scatter==2.0.5
torch-sparse==0.6.6
torch-spline-conv==1.2.0
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.47.0
traitlets==4.3.3
ttach==0.0.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
